requirements.txt
mcp
python-docx
pymupdf
openpyxl
requests